+++ src/azure-cli-core/azure/cli/core/profiles/_shared.py
@@ -26,25 +26,17 @@ PROFILE_TYPE = object()
 
 class ResourceType(Enum):  # pylint: disable=too-few-public-methods
 
-    MGMT_STORAGE = ('azure.mgmt.storage',
-                    'StorageManagementClient')
-    MGMT_COMPUTE = ('azure.mgmt.compute',
-                    'ComputeManagementClient')
-    MGMT_NETWORK = ('azure.mgmt.network',
-                    'NetworkManagementClient')
-    MGMT_RESOURCE_FEATURES = ('azure.mgmt.resource.features',
-                              'FeatureClient')
-    MGMT_RESOURCE_LINKS = ('azure.mgmt.resource.links',
-                           'ManagementLinkClient')
-    MGMT_RESOURCE_LOCKS = ('azure.mgmt.resource.locks',
-                           'ManagementLockClient')
-    MGMT_RESOURCE_POLICY = ('azure.mgmt.resource.policy',
-                            'PolicyClient')
-    MGMT_RESOURCE_RESOURCES = ('azure.mgmt.resource.resources',
-                               'ResourceManagementClient')
-    MGMT_RESOURCE_SUBSCRIPTIONS = ('azure.mgmt.resource.subscriptions',
-                                   'SubscriptionClient')
+    MGMT_STORAGE = ('azure.mgmt.storage', 'StorageManagementClient')
+    MGMT_COMPUTE = ('azure.mgmt.compute', 'ComputeManagementClient')
+    MGMT_NETWORK = ('azure.mgmt.network', 'NetworkManagementClient')
+    MGMT_RESOURCE_FEATURES = ('azure.mgmt.resource.features', 'FeatureClient')
+    MGMT_RESOURCE_LINKS = ('azure.mgmt.resource.links', 'ManagementLinkClient')
+    MGMT_RESOURCE_LOCKS = ('azure.mgmt.resource.locks', 'ManagementLockClient')
+    MGMT_RESOURCE_POLICY = ('azure.mgmt.resource.policy', 'PolicyClient')
+    MGMT_RESOURCE_RESOURCES = ('azure.mgmt.resource.resources', 'ResourceManagementClient')
+    MGMT_RESOURCE_SUBSCRIPTIONS = ('azure.mgmt.resource.subscriptions', 'SubscriptionClient')
     DATA_STORAGE = ('azure.multiapi.storage', None)
+    DATA_COSMOS_TABLE = ('azure.multiapi.cosmosdb', None)
 
     def __init__(self, import_prefix, client_name):
         """Constructor.
@@ -69,7 +61,8 @@ AZURE_API_PROFILES = {
         ResourceType.MGMT_RESOURCE_POLICY: '2017-06-01-preview',
         ResourceType.MGMT_RESOURCE_RESOURCES: '2017-05-10',
         ResourceType.MGMT_RESOURCE_SUBSCRIPTIONS: '2016-06-01',
-        ResourceType.DATA_STORAGE: '2017-04-17'
+        ResourceType.DATA_STORAGE: '2017-04-17',
+        ResourceType.DATA_COSMOS_TABLE: '2017-04-17'
     },
     '2017-03-09-profile': {
         ResourceType.MGMT_STORAGE: '2015-06-15',